# Refactor cmis eeprom i2c read