--- a/sonic_platform_base/sonic_xcvr/api/public/cmis.py
+++ b/sonic_platform_base/sonic_xcvr/api/public/cmis.py
@@ -1,4 +1,3 @@
-
 """
     cmis.py
 
@@ -19,6 +18,7 @@ from .cmisVDM import CmisVdmApi
 import time
 import copy
 from collections import defaultdict
+from ...utils.cache import read_only_cached_api_return
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
@@ -108,32 +108,6 @@ CMIS_XCVR_INFO_DEFAULT_DICT = {
         "specification_compliance": "N/A",
         "vdm_supported": "N/A"
         }
-
-def read_only_cache_none_dict_and_list(func):
-    cache_name = f'_{func.__name__}_cache'
-    def wrapper(self):
-        if not hasattr(self, cache_name):
-            cache_value = func(self)
-            setattr(self, cache_name, cache_value)
-        else:
-            cache_value = getattr(self, cache_name)
-            if cache_value is None:
-                cache_value = func(self)
-                setattr(self, cache_name, cache_value)
-        return cache_value
-    return wrapper
-
-# Decorator that caches return values of dict/list methods only when non-empty collections are returned
-def read_only_cache_dict_and_list(func):
-    cache_name = f'_{func.__name__}_cache'
-    def wrapper(self):
-        if not hasattr(self, cache_name):
-            cache_value = func(self)
-            if cache_value is not None and isinstance(cache_value, (dict, list)) and len(cache_value) > 0:
-                setattr(self, cache_name, cache_value)
-            return cache_value
-        return getattr(self, cache_name)
-    return wrapper
 
 class CmisApi(XcvrApi):
     NUM_CHANNELS = 8
@@ -225,14 +199,14 @@ class CmisApi(XcvrApi):
         '''
         return self.xcvr_eeprom.read(consts.VDM_UNFREEZE_DONE)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_manufacturer(self):
         '''
         This function returns the manufacturer of the module
         '''
         return self.xcvr_eeprom.read(consts.VENDOR_NAME_FIELD)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_model(self):
         '''
         This function returns the part number of the module
@@ -245,49 +219,49 @@ class CmisApi(XcvrApi):
         '''
         return "Length Cable Assembly(m)"
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_cable_length(self):
         '''
         This function returns the cable length of the module
         '''
         return self.xcvr_eeprom.read(consts.LENGTH_ASSEMBLY_FIELD)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_vendor_rev(self):
         '''
         This function returns the revision level for part number provided by vendor
         '''
         return self.xcvr_eeprom.read(consts.VENDOR_REV_FIELD)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_serial(self):
         '''
         This function returns the serial number of the module
         '''
         return self.xcvr_eeprom.read(consts.VENDOR_SERIAL_NO_FIELD)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_module_type(self):
         '''
         This function returns the SFF8024Identifier (module type / form-factor). Table 4-1 in SFF-8024 Rev4.6
         '''
         return self.xcvr_eeprom.read(consts.ID_FIELD)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_module_type_abbreviation(self):
         '''
         This function returns the SFF8024Identifier (module type / form-factor). Table 4-1 in SFF-8024 Rev4.6
         '''
         return self.xcvr_eeprom.read(consts.ID_ABBRV_FIELD)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_connector_type(self):
         '''
         This function returns module connector. Table 4-3 in SFF-8024 Rev4.6
         '''
         return self.xcvr_eeprom.read(consts.CONNECTOR_FIELD)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_module_hardware_revision(self):
         '''
         This function returns the module hardware revision
@@ -299,7 +273,7 @@ class CmisApi(XcvrApi):
         hw_rev = [str(num) for num in [hw_major_rev, hw_minor_rev]]
         return '.'.join(hw_rev)
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_cmis_rev(self):
         '''
         This function returns the CMIS version the module complies to
@@ -614,7 +588,7 @@ class CmisApi(XcvrApi):
         media_intf = self.get_module_media_type()
         return media_intf == "passive_copper_media_interface" if media_intf else None
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def is_flat_memory(self):
         return self.xcvr_eeprom.read(consts.FLAT_MEM_FIELD) is not False
 
@@ -886,7 +860,7 @@ class CmisApi(XcvrApi):
             tx_los_final.append(bool(tx_los[key]))
         return tx_los_final
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_tx_disable_support(self):
         return not self.is_flat_memory() and self.xcvr_eeprom.read(consts.TX_DISABLE_SUPPORT_FIELD)
 
@@ -1000,7 +974,7 @@ class CmisApi(XcvrApi):
     def get_transceiver_thresholds_support(self):
         return not self.is_flat_memory()
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_lpmode_support(self):
         power_class = self.xcvr_eeprom.read(consts.POWER_CLASS_FIELD)
         if power_class is None:
@@ -1042,7 +1016,7 @@ class CmisApi(XcvrApi):
         else:
             return 'Unknown media interface'
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def is_coherent_module(self):
         '''
         Returns True if the module follow C-CMIS spec, False otherwise
@@ -1050,7 +1024,7 @@ class CmisApi(XcvrApi):
         mintf = self.get_module_media_interface()
         return False if 'ZR' not in mintf else True
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_datapath_init_duration(self):
         '''
         This function returns the duration of datapath init
@@ -1063,7 +1037,7 @@ class CmisApi(XcvrApi):
         value = float(duration)
         return value * DATAPATH_INIT_DURATION_MULTIPLIER if value <= DATAPATH_INIT_DURATION_OVERRIDE_THRESHOLD else value
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_datapath_deinit_duration(self):
         '''
         This function returns the duration of datapath deinit
@@ -1073,7 +1047,7 @@ class CmisApi(XcvrApi):
         duration = self.xcvr_eeprom.read(consts.DP_PATH_DEINIT_DURATION)
         return float(duration) if duration is not None else 0
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_datapath_tx_turnon_duration(self):
         '''
         This function returns the duration of datapath tx turnon
@@ -1083,7 +1057,7 @@ class CmisApi(XcvrApi):
         duration = self.xcvr_eeprom.read(consts.DP_TX_TURNON_DURATION)
         return float(duration) if duration is not None else 0
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_datapath_tx_turnoff_duration(self):
         '''
         This function returns the duration of datapath tx turnoff
@@ -1093,7 +1067,7 @@ class CmisApi(XcvrApi):
         duration = self.xcvr_eeprom.read(consts.DP_TX_TURNOFF_DURATION)
         return float(duration) if duration is not None else 0
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_module_pwr_up_duration(self):
         '''
         This function returns the duration of module power up
@@ -1103,7 +1077,7 @@ class CmisApi(XcvrApi):
         duration = self.xcvr_eeprom.read(consts.MODULE_PWRUP_DURATION)
         return float(duration) if duration is not None else 0
 
-    @read_only_cache_none_dict_and_list
+    @read_only_cached_api_return
     def get_module_pwr_down_duration(self):
         '''
         This function returns the duration of module power down
@@ -2805,7 +2779,7 @@ class CmisApi(XcvrApi):
             return None
         return [bool(datapath_deinit & (1 << lane)) for lane in range(self.NUM_CHANNELS)]
 
-    @read_only_cache_dict_and_list
+    @read_only_cached_api_return
     def get_application_advertisement(self):
         """
         Get the application advertisement of the CMIS transceiver